refactor(harness)!: consolidate the Pydantic-AI harness, remove tracing handler
Collapse _pydantic_ai_async / _pydantic_ai_tracing into _pydantic_ai_sync
(convert helper) and _pydantic_ai_turn (stream_pydantic_ai_events). Span
tracing is derived from the canonical stream by UnifiedEmitter, so
create_pydantic_ai_tracing_handler is removed.

BREAKING CHANGE: create_pydantic_ai_tracing_handler is removed.
