+++ src/agentex/lib/adk/__init__.py
@@ -11,10 +11,8 @@ from agentex.lib.adk._modules._langgraph_sync import (
     emit_langgraph_messages,
     convert_langgraph_to_agentex_events,
 )
-from agentex.lib.adk._modules._pydantic_ai_async import stream_pydantic_ai_events
+from agentex.lib.adk._modules._pydantic_ai_turn import PydanticAITurn, stream_pydantic_ai_events
 from agentex.lib.adk._modules._pydantic_ai_sync import convert_pydantic_ai_to_agentex_events
-from agentex.lib.adk._modules._pydantic_ai_tracing import create_pydantic_ai_tracing_handler
-from agentex.lib.adk._modules._pydantic_ai_turn import PydanticAITurn
 from agentex.lib.adk._modules._claude_code_sync import convert_claude_code_to_agentex_events
 from agentex.lib.adk._modules._claude_code_turn import (
     ClaudeCodeTurn,
@@ -75,7 +73,6 @@ __all__ = [
     # Pydantic AI
     "stream_pydantic_ai_events",
     "convert_pydantic_ai_to_agentex_events",
-    "create_pydantic_ai_tracing_handler",
     "PydanticAITurn",
     # Claude Code
     "convert_claude_code_to_agentex_events",
