+++ tests/lib/adk/test_pydantic_ai_sync.py
@@ -1,4 +1,12 @@
-"""Tests for the Pydantic AI -> Agentex stream event converter."""
+"""Tests for the sync Pydantic AI -> Agentex path.
+
+Covers:
+- The bare converter ``convert_pydantic_ai_to_agentex_events`` (text/thinking/
+  tool-call streaming and arg-delta handling).
+- The unified sync (HTTP ACP) path ``UnifiedEmitter.yield_turn(PydanticAITurn(...))``:
+    * Passthrough: yield_turn events equal PydanticAITurn(stream).events
+    * Span derivation (tool + reasoning) with a fake tracing backend
+"""
 
 from __future__ import annotations
 
@@ -25,6 +33,7 @@ from pydantic_ai.messages import (
     FunctionToolResultEvent,
 )
 
+from agentex.lib.core.harness import UnifiedEmitter
 from agentex.types.reasoning_content import ReasoningContent
 from agentex.types.task_message_delta import TextDelta
 from agentex.types.tool_request_delta import ToolRequestDelta
@@ -42,6 +51,9 @@ from agentex.lib.adk._modules._pydantic_ai_sync import (
     _args_delta_to_str,
     convert_pydantic_ai_to_agentex_events,
 )
+from agentex.lib.adk._modules._pydantic_ai_turn import PydanticAITurn
+
+from ..core.harness._fakes import FakeTracing
 
 
 async def _aiter(events: list[Any]) -> AsyncIterator[Any]:
@@ -290,90 +302,6 @@ class TestToolCallStreaming:
         assert out[0].content.content == "bad arguments"
 
 
-class TestTracingHandlerSync:
-    """The sync converter has the same opt-in tracing-handler contract as the
-    async streamer: pass a handler and the converter calls ``on_tool_start`` /
-    ``on_tool_end`` for each tool call. Streaming yields are unchanged when
-    omitted."""
-
-    class _RecordingHandler:
-        def __init__(self) -> None:
-            self.starts: list[dict[str, Any]] = []
-            self.ends: list[dict[str, Any]] = []
-
-        async def on_tool_start(self, tool_call_id: str, tool_name: str, arguments: Any) -> None:
-            self.starts.append({"tool_call_id": tool_call_id, "tool_name": tool_name, "arguments": arguments})
-
-        async def on_tool_end(self, tool_call_id: str, result: Any) -> None:
-            self.ends.append({"tool_call_id": tool_call_id, "result": result})
-
-    async def test_handler_records_start_and_end_for_a_tool_call(self):
-        handler = self._RecordingHandler()
-        events = [
-            PartStartEvent(
-                index=0,
-                part=ToolCallPart(tool_name="get_weather", args=None, tool_call_id="c1"),
-            ),
-            PartEndEvent(
-                index=0,
-                part=ToolCallPart(tool_name="get_weather", args='{"city":"Paris"}', tool_call_id="c1"),
-            ),
-            FunctionToolResultEvent(
-                part=ToolReturnPart(tool_name="get_weather", content="Sunny", tool_call_id="c1"),
-            ),
-        ]
-        out = await _collect(
-            convert_pydantic_ai_to_agentex_events(_aiter(events), tracing_handler=handler)  # type: ignore[arg-type]
-        )
-
-        # Streaming output is unchanged.
-        assert any(isinstance(e, StreamTaskMessageStart) for e in out)
-        assert any(isinstance(e, StreamTaskMessageFull) for e in out)
-
-        assert handler.starts == [
-            {
-                "tool_call_id": "c1",
-                "tool_name": "get_weather",
-                "arguments": {"city": "Paris"},
-            }
-        ]
-        assert handler.ends == [{"tool_call_id": "c1", "result": "Sunny"}]
-
-    async def test_handler_not_called_when_no_tool_calls(self):
-        handler = self._RecordingHandler()
-        events = [
-            PartStartEvent(index=0, part=TextPart(content="")),
-            PartDeltaEvent(index=0, delta=TextPartDelta(content_delta="hi")),
-            PartEndEvent(index=0, part=TextPart(content="hi")),
-        ]
-        await _collect(
-            convert_pydantic_ai_to_agentex_events(_aiter(events), tracing_handler=handler)  # type: ignore[arg-type]
-        )
-        assert handler.starts == []
-        assert handler.ends == []
-
-    async def test_omitting_handler_preserves_pre_tracing_behavior(self):
-        events = [
-            PartStartEvent(
-                index=0,
-                part=ToolCallPart(tool_name="t", args=None, tool_call_id="c"),
-            ),
-            PartEndEvent(
-                index=0,
-                part=ToolCallPart(tool_name="t", args="{}", tool_call_id="c"),
-            ),
-            FunctionToolResultEvent(
-                part=ToolReturnPart(tool_name="t", content="ok", tool_call_id="c"),
-            ),
-        ]
-        out = await _collect(convert_pydantic_ai_to_agentex_events(_aiter(events)))
-        # Same emit shape as before: Start, Done, Full
-        types = [type(e).__name__ for e in out]
-        assert "StreamTaskMessageStart" in types
-        assert "StreamTaskMessageDone" in types
-        assert "StreamTaskMessageFull" in types
-
-
 class TestMultiStepRun:
     async def test_text_then_tool_then_text_assigns_distinct_indices(self):
         """A multi-step run: model emits text + tool call → tool runs → model emits more text.
@@ -555,3 +483,157 @@ class TestOnResultCallback:
 
         assert len(awaited) == 1
         assert awaited[0].result.output == "async_output"
+
+
+# ---------------------------------------------------------------------------
+# Unified sync path: PydanticAITurn + UnifiedEmitter.yield_turn
+#
+# Exercises the path documented in _pydantic_ai_sync.py under
+# "Recommended: unified surface":
+# - events forwarded by yield_turn equal PydanticAITurn(stream).events (passthrough)
+# - with a trace context + fake tracing backend, tool / reasoning spans are derived
+# ---------------------------------------------------------------------------
+
+
+class TestUnifiedSyncPathPassthrough:
+    """The events forwarded by yield_turn are identical to PydanticAITurn.events."""
+
+    async def test_text_stream_passthrough(self):
+        raw_events = [
+            PartStartEvent(index=0, part=TextPart(content="")),
+            PartDeltaEvent(index=0, delta=TextPartDelta(content_delta="hello")),
+            PartEndEvent(index=0, part=TextPart(content="hello")),
+        ]
+
+        turn_a = PydanticAITurn(_aiter(raw_events), model="openai:gpt-4o")
+        direct = await _collect(turn_a.events)
+
+        turn_b = PydanticAITurn(_aiter(raw_events), model="openai:gpt-4o")
+        emitter = UnifiedEmitter(task_id="t", trace_id=None, parent_span_id=None)
+        via_emitter = await _collect(emitter.yield_turn(turn_b))
+
+        assert len(via_emitter) == len(direct)
+        for a, b in zip(via_emitter, direct):
+            assert type(a) is type(b)
+            assert a.model_dump() == b.model_dump()
+
+    async def test_tool_call_stream_passthrough(self):
+        raw_events = [
+            PartStartEvent(index=0, part=ToolCallPart(tool_name="Bash", args=None, tool_call_id="c1")),
+            PartDeltaEvent(index=0, delta=ToolCallPartDelta(args_delta='{"cmd":"ls"}')),
+            PartEndEvent(
+                index=0,
+                part=ToolCallPart(tool_name="Bash", args='{"cmd":"ls"}', tool_call_id="c1"),
+            ),
+        ]
+
+        turn_a = PydanticAITurn(_aiter(raw_events), model="openai:gpt-4o")
+        direct = await _collect(turn_a.events)
+
+        turn_b = PydanticAITurn(_aiter(raw_events), model="openai:gpt-4o")
+        emitter = UnifiedEmitter(task_id="t", trace_id=None, parent_span_id=None)
+        via_emitter = await _collect(emitter.yield_turn(turn_b))
+
+        assert len(via_emitter) == len(direct)
+        for a, b in zip(via_emitter, direct):
+            assert type(a) is type(b)
+            assert a.model_dump() == b.model_dump()
+
+
+class TestUnifiedSyncPathSpanDerivation:
+    """With trace context + fake tracing, spans are derived from the stream."""
+
+    async def test_tool_span_opened_and_closed(self):
+        """A tool call produces start_span + end_span on the fake tracing backend."""
+        tool_events = [
+            PartStartEvent(
+                index=0,
+                part=ToolCallPart(tool_name="Bash", args={"cmd": "ls"}, tool_call_id="call_1"),
+            ),
+            PartEndEvent(
+                index=0,
+                part=ToolCallPart(tool_name="Bash", args='{"cmd":"ls"}', tool_call_id="call_1"),
+            ),
+            FunctionToolResultEvent(
+                part=ToolReturnPart(tool_name="Bash", content="files", tool_call_id="call_1"),
+            ),
+        ]
+
+        fake = FakeTracing()
+        turn = PydanticAITurn(_aiter(tool_events), model="openai:gpt-4o")
+        emitter = UnifiedEmitter(task_id="t", trace_id="tr", parent_span_id="p", tracing=fake)
+
+        events = await _collect(emitter.yield_turn(turn))
+
+        assert len(events) >= 2, "at least Start(tool) + Done + Full(response)"
+        assert len(fake.started) == 1, "one tool span opened"
+        assert len(fake.ended) == 1, "one tool span closed"
+        span_name, parent_id, span_input = fake.started[0]
+        assert span_name == "Bash"
+        assert parent_id == "p"
+        closed_name, closed_output = fake.ended[0]
+        assert closed_name == "Bash"
+
+    async def test_reasoning_span_opened_and_closed(self):
+        """A thinking/reasoning block produces start_span + end_span."""
+        reasoning_events = [
+            PartStartEvent(index=0, part=ThinkingPart(content="")),
+            PartDeltaEvent(index=0, delta=ThinkingPartDelta(content_delta="let me think")),
+            PartEndEvent(index=0, part=ThinkingPart(content="let me think")),
+        ]
+
+        fake = FakeTracing()
+        turn = PydanticAITurn(_aiter(reasoning_events), model="openai:gpt-4o")
+        emitter = UnifiedEmitter(task_id="t", trace_id="tr", parent_span_id="p", tracing=fake)
+
+        await _collect(emitter.yield_turn(turn))
+
+        assert len(fake.started) == 1, "one reasoning span opened"
+        assert len(fake.ended) == 1, "one reasoning span closed"
+        span_name, parent_id, _ = fake.started[0]
+        assert span_name == "reasoning"
+        assert parent_id == "p"
+
+    async def test_no_trace_id_means_no_spans(self):
+        """When trace_id is None, no spans are derived even with a fake tracing backend."""
+        raw_events = [
+            PartStartEvent(
+                index=0,
+                part=ToolCallPart(tool_name="Bash", args={"cmd": "ls"}, tool_call_id="c2"),
+            ),
+            PartEndEvent(
+                index=0,
+                part=ToolCallPart(tool_name="Bash", args='{"cmd":"ls"}', tool_call_id="c2"),
+            ),
+        ]
+
+        fake = FakeTracing()
+        turn = PydanticAITurn(_aiter(raw_events), model="openai:gpt-4o")
+        emitter = UnifiedEmitter(task_id="t", trace_id=None, parent_span_id=None, tracing=fake)
+
+        await _collect(emitter.yield_turn(turn))
+
+        assert fake.started == [], "no spans when trace_id is absent"
+        assert fake.ended == []
+
+    async def test_tracer_false_suppresses_spans_even_with_trace_id(self):
+        """tracer=False disables span derivation regardless of trace_id."""
+        raw_events = [
+            PartStartEvent(
+                index=0,
+                part=ToolCallPart(tool_name="Bash", args={"cmd": "ls"}, tool_call_id="c3"),
+            ),
+            PartEndEvent(
+                index=0,
+                part=ToolCallPart(tool_name="Bash", args='{"cmd":"ls"}', tool_call_id="c3"),
+            ),
+        ]
+
+        fake = FakeTracing()
+        turn = PydanticAITurn(_aiter(raw_events), model="openai:gpt-4o")
+        emitter = UnifiedEmitter(task_id="t", trace_id="tr", parent_span_id="p", tracer=False, tracing=fake)
+
+        await _collect(emitter.yield_turn(turn))
+
+        assert fake.started == []
+        assert fake.ended == []
